Refactor User Meta endpoint integration tests to reduce code duplication
